Refactor data models and mappers for better modularity and separation of concerns.
- Introduced separate mapper files for each domain entity (Cart, Category, Product, User) to improve modularity and maintainability.
- Removed the monolithic `mapper.dart` file and distributed its functionality across individual mapper files.
- Updated repository implementations to reference the new mapper files.
- Added a new `UserEntity` class in the domain layer to represent user data.
- Updated the `AuthEntity` to include the new `UserEntity` and refactored related files accordingly.
- Ensured all extensions and mappers follow the naming convention for better readability and consistency.

This refactor aims to enhance the clarity and maintainability of the codebase by clearly separating concerns and improving the structure of data mapping logic.
